Clean up stale refactor references
